+++ packages/infra/test/requires.test.ts
@@ -1,6 +1,7 @@
-import { Effect } from "effect-app"
+import { it } from "@effect/vitest"
+import { Effect, S } from "effect-app"
 import { Middleware } from "../src/api/routing.js"
-import { SomeService } from "./query.test.js"
+import { AllowAnonymous, type RequestContextMap, RequireRoles, SomeService, Test } from "./fixtures.js"
 import { makeNewMiddleware } from "./requires.js"
 
 export class SomeMiddleware extends Middleware.Tag<SomeMiddleware>()("SomeMiddleware", {
@@ -12,7 +13,7 @@ export class SomeMiddleware extends Middleware.Tag<SomeMiddleware>()("SomeMiddle
     return ({ next }) =>
       Effect.gen(function*() {
         // yield* Effect.context<"test-dep2">()
-        return yield* next.pipe(Effect.provideService(SomeService, null as any))
+        return yield* next.pipe(Effect.provideService(SomeService, new SomeService({ a: 1 })))
       })
   })
 }) {
@@ -34,6 +35,19 @@ export class RequiresSomeMiddleware extends Middleware.Tag<RequiresSomeMiddlewar
 }) {
 }
 
-export const middleware3 = makeNewMiddleware<{}>()()
-  .middleware(RequiresSomeMiddleware)
-  .middleware(SomeMiddleware)
+it("requires gets enforced", async () => {
+  const middleware3 = makeNewMiddleware<RequestContextMap>()
+    .middleware(RequiresSomeMiddleware)
+    .middleware(SomeMiddleware)
+    .addDynamicMiddleware(AllowAnonymous)
+    .addDynamicMiddleware(RequireRoles)
+    .addDynamicMiddleware(Test)
+
+  await Effect
+    .gen(function*() {
+      const mw = yield* middleware3
+      const mwM = mw.effect(Object.assign({}, S.Any, { config: {} }), (req) => Effect.void, "some-module")
+      const v = yield* mwM({}, {})
+    })
+    .pipe(Effect.provide(middleware3.Default), Effect.runPromise)
+})
